feat: Updates according to feedback from Andy (GitHub user ndp)
